Apr 6, 2025, 3:24 AM

+++ CT06_11/lesson11.py
@@ -17,6 +17,6 @@
 # else:
 #     print("X")
 
-num = int(input("Choose a random number"))
+num = float(input("Choose a random number"))
 result= num/3%0 and num/7%0
 print(str(result))
